// angular/src/app/model/feed-model.ts
export interface FeedModel {
  FLSR_ID: string
  FSIN_ID: string
  ID: string
  PRICE: string
  FLY_DURATION: string
  RET_DURATION: string
  TOTAL_DURATION: string
  ID_OUT: string
  DEP_DATE_OUT: string
  ARR_DATE_OUT: string
  AIRLINE_OUT: string
  COUNTRY_FROM_OUT: string
  COUNTRY_TO_OUT: string
  CITY_FROM_OUT: string
  CITY_TO_OUT: string
  FLIGHT_NO_OUT: string
  ID_BACK: string
  DEP_DATE_BACK: string
  ARR_DATE_BACK: string
  AIRLINE_BACK: string
  COUNTRY_FROM_BACK: string
  COUNTRY_TO_BACK: string
  CITY_FROM_BACK: string
  CITY_TO_BACK: string
  FLIGHT_NO_BACK: string
  FX_RATE: string
  EUR_PRICE: string
  CR_DATE: string
  SEC_PRICE: string
  TH_PRICE: string
  FO_PRICE: string
  }